+++ signin.html
@@ -51,6 +51,6 @@
       </div>
     </div> 
   </div>
-  <script type="module" src="./signin.js"></script>
+  <script type="module" src="./javascript/signin.js"></script>
 </body>
 </html>
