+++ signup.html
@@ -59,6 +59,6 @@
       </div>
     </div> 
   </div>
-  <script type="module" src="./signup.js"></script>
+  <script type="module" src="./javascript/signup.js"></script>
 </body>
 </html>
